Improved gauge + click handling
